Get instance of PGRM

This function generates a PGRM copy with specified ancestry, build, and phecode version

Usage

get_PGRM(
  ancestry = "all",
  build = "hg19",
  phecode_version = "V1.2",
  unique = TRUE
)

Arguments

ancestry

A string that indicates the ancestry of the PGRM. Options EAS, EUR, AFR, SAS, AMR, ALL. Default ALL

build

A string indicating the genome reference build. Options hg19, hg38. Default is hg19.

phecode_version

A string indicating the phecode version. Currently only V1.2 is supported, which is the default

unique

If TRUE, then rows are uniqued by SNP/phecode (only relevant when ancestry == "ALL")

Value

A data.table of the PGRM

Details

This function returns a copy of the PGRM with specified columns for SNP, AF, and cases_needed according to the arguments specified. The function assigns a column "SNP" as either SNP_hg19 or SNP_hg38, depending on the build argument. It also assigns the allele frequency (AF) and cases_needed columns according to the ancestry. If the ancestry is != "ALL", then the funciton also subsets PGRM_ALL to include only associations that were based in cohorts with the specified ancestry. This funciton is called inside annotate_results(), which can be used to annoate the results of a specific test cohort.

See also

Examples

library(pgrm)

## Get a copy of the PGRM for build hg19, East Asian ancestry
get_PGRM(build='hg19',ancestry='EAS')
#>       assoc_ID              SNP ancestry        rsID risk_allele_dir
#>    1:    59717 10:101290301:T:C      EAS   rs6584283             alt
#>    2:   247636 10:104456686:T:C      EAS  rs10883765             alt
#>    3:   103387 10:104595248:G:A      EAS   rs4919687             ref
#>    4:   247628 10:104654577:T:C      EAS  rs10883795             alt
#>    5:   216343 10:104719096:G:A      EAS  rs12413409             ref
#>   ---                                                               
#> 1211:   112945   9:84308948:G:A      EAS   rs2796441             ref
#> 1212:   180392    9:8879118:C:T      EAS  rs17584499             alt
#> 1213:   216979   9:92202495:T:C      EAS  rs10797119             alt
#> 1214:    84289   9:93636664:G:A      EAS    rs290227             ref
#> 1215:   112946   9:98278413:C:T      EAS rs113154802             ref
#>       risk_allele     AF phecode    phecode_string    category_string
#>    1:           C 0.5569     153 Colorectal cancer          neoplasms
#>    2:           C 0.4108   295.1     Schizophrenia        psychiatric
#>    3:           G 0.2494     153 Colorectal cancer          neoplasms
#>    4:           C 0.5233   295.1     Schizophrenia        psychiatric
#>    5:           G 0.2816     401      Hypertension circulatory system
#>   ---                                                                
#> 1211:           G 0.6390   250.2   Type 2 diabetes          endocrine
#> 1212:           T 0.1050   250.2   Type 2 diabetes          endocrine
#> 1213:           C 0.2571     495            Asthma        respiratory
#> 1214:           G 0.5868  290.16 Vascular dementia       neurological
#> 1215:           C 0.1071   250.2   Type 2 diabetes          endocrine
#>       cat_LOG10_P   cat_OR  cat_L95  cat_U95 cases_needed Study_accession
#>    1:   10.221849 1.086957 1.060000 1.110000         5873      GCST009869
#>    2:    8.522879 1.086957 1.040000 1.130000        13139      GCST003880
#>    3:   11.096910 1.140000 1.100000 1.190000         2943      GCST003799
#>    4:    9.096910 1.150000 1.100000 1.200000         2171      GCST003880
#>    5:    8.301030 1.176471 1.111111 1.234568         2228      GCST011141
#>   ---                                                                    
#> 1211:   28.000000 1.080000 1.060000 1.090000         6225      GCST010118
#> 1212:    9.045757 1.570000 1.360000 1.820000          506      GCST000601
#> 1213:    8.000000 1.107381 1.068997 1.147144         5698    GCST90013679
#> 1214:   10.154902 2.200000 1.730000 2.780000           67      GCST000993
#> 1215:    7.397940 1.060000 1.040000 1.080000        33750      GCST010118