Calculate the replicaiton rate (RR) of a test cohort

This function calculates the replicaiton rate in a test cohort that has been annotated with PGRM. By default, it calculates the RR for associations that powered at >80%.

Usage

get_RR(annotated_results, include = "powered", LOUD = TRUE)

Arguments

annotated_results

A data table of results that have been annotated with the PGRM

include

A character string. If 'powered' then only powered associations are included. If 'all' then all associations are included

LOUD

If TRUE then progress info is printed to the terminal. Default TRUE

Value

An numeric value of the replication rate of the result set

Examples

  library(pgrm)
  library(data.table)

  ## Read in an ICD file. Format should look like the example file.
  ## Note: Don't forget to load icd column as a character vector!
  head(icdExampleTable)
#>    person_id    icd flag entry_date
#> 1:         1  034.0    9 2014-04-12
#> 2:         3  041.9    9 2014-08-06
#> 3:         1 079.99    9 2011-05-30
#> 4:         3  112.1    9 2014-08-06
#> 5:         2  150.5    9 2014-02-17
#> 6:         2  150.5    9 2014-04-18

  ## make pheno table
  pheno = make_pheno(icdExampleTable)
  head(pheno)
#>    person_id phecode N
#> 1:         1     465 5
#> 2:         3     041 2
#> 3:         1     079 2
#> 4:         3     112 1
#> 5:         2     150 3
#> 6:         2     151 6

  ## Load in covariate file. Here's a generated example
  covar_table_test = data.table(
  person_id = 1:4, sex = c('F', 'M', 'F', 'M'),
  last_age = c(28772, 18028, 11636, 14589))

  ## Load genotype file
  library(gaston)
#> Loading required package: Rcpp
#> Loading required package: RcppParallel
#> 
#> Attaching package: ‘RcppParallel’
#> The following object is masked from ‘package:Rcpp’:
#> 
#>     LdFlags
#> Gaston set number of threads to 1. Use setThreadOptions() to modify this.
#> 
#> Attaching package: ‘gaston’
#> The following object is masked from ‘package:stats’:
#> 
#>     sigma
#> The following objects are masked from ‘package:base’:
#> 
#>     cbind, rbind
 # geno = read.bed.matrix('data/geno_test')

  ## Make a PGRM instance, specifying genome build and ancestry
  PGRM = get_PGRM(build = 'hg19', ancestry = 'all')

  ## Run associations from the PGRM
  run_PGRM_assoc(geno,pheno,covars,covariate_names=c('last_age'),PGRM,MCC=2,use_exclude_range=TRUE,check_sex=TRUE)
#> Error in checkMultiClass(x, classes, null.ok): object 'geno' not found