Run associations in the pgrm on data from a test cohort. Function requires row level phenotype, genotype, and covariate information.

(NOTE: If you used principle components as covariates, make sure they are not coded as numeric. Otherwise the logistic regression will treat them as categorical variables, making the function run very slowly before failing.)

Usage

run_PGRM_assoc(
  geno,
  pheno,
  covars,
  covariate_names,
  PGRM,
  MCC = 2,
  minimum_case_count = 100,
  use_exclude_ranges = TRUE,
  check_sex = FALSE,
  LOUD = TRUE
)

Arguments

geno

A matrix or 'BEDMatrix' object containing genetic data

pheno

A data.table with phecode phenotypes. columns person_id, phecode, and N required. The phecode columns should be a character vector. Column N specifies the number of unique dates a phecode occurs for that individual. Only phecodes that are present for an individual are included (i.e. N is always greater than one).

covars

A data.table having one row per person in the cohort. Must havea column names person_id. Must include other columns specified in covariate_names.

covariate_names

A list of covariates to be used in the logistic regression. All covariates listed must be present in the covars table

PGRM

A data.table of the PGRM, generated with get_PGRM()

MCC

An integer specifying the minimum code count required for a case (default MCC=2)

minimum_case_count

An integer specifiying the minimum number of cases required in the test cohort to be included in the analysis (default minimum_case_count=100)

use_exclude_ranges

If TRUE then exclude ranges are applied to controls

check_sex

If TRUE then individuals with sex == F in the covars table will be excluded from male-specific phenotypes, and vice-versa. Sex specific phecodes are specified in phecode_info. For this function to work, the covars table must have column sex and the values must be 'M' for Male and 'F' for Female

LOUD

If TRUE then progress info is printed to the terminal. Default TRUE

Value

A data.table with annotated results from association tests

Examples

  library(pgrm)
  library(data.table)

  ## Read in an ICD file. Format should look like the example file.
  ## Note: Don't forget to load icd column as a character vector!
  head(icdExampleTable)
#>    person_id    icd flag entry_date
#> 1:         1  034.0    9 2014-04-12
#> 2:         3  041.9    9 2014-08-06
#> 3:         1 079.99    9 2011-05-30
#> 4:         3  112.1    9 2014-08-06
#> 5:         2  150.5    9 2014-02-17
#> 6:         2  150.5    9 2014-04-18

  ## make pheno table
  pheno = make_pheno(icdExampleTable)
  head(pheno)
#>    person_id phecode N
#> 1:         1     465 5
#> 2:         3     041 2
#> 3:         1     079 2
#> 4:         3     112 1
#> 5:         2     150 3
#> 6:         2     151 6

  ## Load in covariate file. Here's a generated example
  covar_table_test = data.table(
  person_id = 1:4, sex = c('F', 'M', 'F', 'M'),
  last_age = c(28772, 18028, 11636, 14589))

  ## Load genotype file
  library(gaston)
 # geno = read.bed.matrix('data/geno_test')

  ## Make a PGRM instance, specifying genome build and ancestry
  PGRM = get_PGRM(build = 'hg19', ancestry = 'all')

  ## Run associations from the PGRM
 #run_PGRM_assoc(geno,pheno,covars,covariate_names=c('last_age'),PGRM,MCC=2,use_exclude_range=TRUE,check_sex=TRUE)