Get instance of PGRM

This function generates a PGRM copy with specified ancestry, build, and phecode version

Usage

get_PGRM(
  ancestry = "all",
  build = "hg19",
  phecode_version = "V1.2",
  unique = TRUE
)

Arguments

ancestry

A string that indicates the ancestry of the PGRM. Options EAS, EUR, AFR, SAS, AMR, ALL. Default ALL

build

A string indicating the genome reference build. Options hg19, hg38. Default is hg19.

phecode_version

A string indicating the phecode version. Currently only V1.2 is supported, which is the default

unique

If TRUE, then rows are uniqued by SNP/phecode (only relevant when ancestry == 'ALL')

Value

A data.table of the PGRM

Details

This function returns a copy of the PGRM. The function assigns a column 'SNP' as either SNP_hg19 or SNP_hg38, depending on the build argument. The returned PGRM only includes associations annotated with the specified ancestry, and the risk allele frequency (RAF) is based on the specified ancestry.

See also

Examples

library(pgrm)

## Get a copy of the PGRM for build hg19, East Asian ancestry
get_PGRM(build = 'hg19', ancestry = 'EAS')
#>       assoc_ID             SNP ancestry        rsID risk_allele_dir        RAF
#>    1:     7883 4:174444902:G:A      EAS  rs17059534             alt 0.07270002
#>    2:     7884 4:111718067:G:A      EAS   rs6843082             ref 0.75449997
#>    3:     7885 4:111761487:T:C      EAS   rs3853445             ref 0.63690001
#>    4:     7886 1:112392360:G:T      EAS  rs12044963             alt 0.34810001
#>    5:     7887 1:203026591:A:G      EAS  rs17461925             ref 0.80930001
#>   ---                                                                         
#> 1211:   306221  6:32575513:C:G      EAS   rs9271060             alt 0.81039999
#> 1212:   306222  6:83648056:C:T      EAS    rs293526             alt 0.79761900
#> 1213:   306223 9:117568766:A:G      EAS   rs6478109             alt 0.47939998
#> 1214:   306224  6:32659443:A:C      EAS rs140068907             alt 0.06639999
#> 1215:   311954  18:6282593:A:C      EAS    rs403814             ref 0.75610000
#>       phecode      phecode_string    category_string cat_LOG10_P   cat_OR
#>    1:  427.21 Atrial fibrillation circulatory system    9.522879 1.201000
#>    2:  427.21 Atrial fibrillation circulatory system  126.000000 1.842000
#>    3:  427.21 Atrial fibrillation circulatory system   33.698970 1.330000
#>    4:  427.21 Atrial fibrillation circulatory system   14.397940 1.110000
#>    5:  427.21 Atrial fibrillation circulatory system   10.045757 1.077000
#>   ---                                                                    
#> 1211:   555.1  Regional enteritis          digestive   22.000000 2.857143
#> 1212:   555.1  Regional enteritis          digestive   10.698970 2.127660
#> 1213:   555.1  Regional enteritis          digestive   24.301030 2.083334
#> 1214:   555.1  Regional enteritis          digestive   11.000000 1.980000
#> 1215:     401        Hypertension circulatory system    8.221849 1.150000
#>       cat_L95 cat_U95 Study_accession pub_count   pub_date
#>    1:   1.134   1.271      GCST004373         1 2017-04-17
#>    2:   1.751   1.939      GCST004373         1 2017-04-17
#>    3:   1.271   1.391      GCST004373         1 2017-04-17
#>    4:   1.082   1.140      GCST004373         2 2017-04-17
#>    5:   1.053   1.102      GCST004373         1 2017-04-17
#>   ---                                                     
#> 1211:   2.330   3.570      GCST003360         1 2016-02-18
#> 1212:   1.690   2.630      GCST003360         1 2016-02-18
#> 1213:   1.820   2.380      GCST003360         1 2016-02-18
#> 1214:   1.620   2.410      GCST003360         1 2016-02-18
#> 1215:   1.070   1.230      GCST003613         1 2016-08-02